+++ src/njs/src/njsPool.cpp
@@ -135,7 +135,7 @@ Handle<Value> Pool::GetPoolMin (Local<String> property,
   if(!njsPool->isValid_)
   {
     string msg = NJSMessages::getErrorMsg(errInvalidPool);
-    NJS_SET_EXCEPTION(msg.c_str(), msg.length());
+    NJS_SET_EXCEPTION(msg.c_str(), (int) msg.length());
     return Undefined();
   }
   else
@@ -159,7 +159,7 @@ Handle<Value> Pool::GetPoolMax (Local<String> property,
   if(!njsPool->isValid_)
   {
     string msg = NJSMessages::getErrorMsg(errInvalidPool);
-    NJS_SET_EXCEPTION(msg.c_str(), msg.length());
+    NJS_SET_EXCEPTION(msg.c_str(), (int) msg.length());
     return Undefined();
   }
   else
@@ -183,7 +183,7 @@ Handle<Value> Pool::GetPoolIncrement (Local<String> property,
   if(!njsPool->isValid_)
   {
     string msg = NJSMessages::getErrorMsg(errInvalidPool);
-    NJS_SET_EXCEPTION(msg.c_str(), msg.length());
+    NJS_SET_EXCEPTION(msg.c_str(), (int) msg.length());
     return Undefined();
   }
   else
@@ -207,7 +207,7 @@ Handle<Value> Pool::GetPoolTimeout (Local<String> property,
   if(!njsPool->isValid_)
   {
     string msg = NJSMessages::getErrorMsg(errInvalidPool);
-    NJS_SET_EXCEPTION(msg.c_str(), msg.length());
+    NJS_SET_EXCEPTION(msg.c_str(), (int) msg.length());
     return Undefined();
   }
   else
@@ -231,7 +231,7 @@ Handle<Value> Pool::GetConnectionsOpen (Local<String> property,
   if(!njsPool->isValid_)
   {
     string error = NJSMessages::getErrorMsg ( errInvalidPool );
-    NJS_SET_EXCEPTION(error.c_str(), error.length());
+    NJS_SET_EXCEPTION(error.c_str(), (int) error.length());
     return scope.Close(Undefined());
   }
   try
@@ -242,7 +242,7 @@ Handle<Value> Pool::GetConnectionsOpen (Local<String> property,
   }
   catch(dpi::Exception &e)
   {
-    NJS_SET_EXCEPTION(e.what(), strlen(e.what()));
+    NJS_SET_EXCEPTION(e.what(), (int) strlen(e.what()));
     return Undefined();
   }
   return Undefined();
@@ -262,7 +262,7 @@ Handle<Value> Pool::GetConnectionsInUse (Local<String> property,
   if(!njsPool->isValid_)
   {
     string error = NJSMessages::getErrorMsg ( errInvalidPool );
-    NJS_SET_EXCEPTION(error.c_str(), error.length());
+    NJS_SET_EXCEPTION(error.c_str(), (int) error.length());
     return scope.Close(Undefined());
   }
   try
@@ -273,7 +273,7 @@ Handle<Value> Pool::GetConnectionsInUse (Local<String> property,
   }
   catch(dpi::Exception &e)
   {
-    NJS_SET_EXCEPTION(e.what(), strlen(e.what()));
+    NJS_SET_EXCEPTION(e.what(), (int) strlen(e.what()));
     return Undefined();
   }
   return Undefined();
@@ -292,7 +292,7 @@ Handle<Value> Pool::GetStmtCacheSize (Local<String> property,
   if(!njsPool->isValid_)
   {
     string msg = NJSMessages::getErrorMsg(errInvalidPool);
-    NJS_SET_EXCEPTION(msg.c_str(), msg.length());
+    NJS_SET_EXCEPTION(msg.c_str(), (int) msg.length());
     return Undefined();
   }
   else 
@@ -318,7 +318,7 @@ void Pool::SetPoolMin (Local<String> property, Local<Value> value,
     msg = NJSMessages::getErrorMsg(errInvalidPool);
   else
     msg = NJSMessages::getErrorMsg(errReadOnly, "poolMin");
-  NJS_SET_EXCEPTION(msg.c_str(), msg.length());
+  NJS_SET_EXCEPTION(msg.c_str(), (int) msg.length());
 }
 
 /*****************************************************************************/
@@ -336,7 +336,7 @@ void Pool::SetPoolMax (Local<String> property, Local<Value> value,
     msg = NJSMessages::getErrorMsg(errInvalidPool);
   else
     msg = NJSMessages::getErrorMsg(errReadOnly, "poolMax");
-  NJS_SET_EXCEPTION(msg.c_str(), msg.length());
+  NJS_SET_EXCEPTION(msg.c_str(), (int) msg.length());
 }
 
 /*****************************************************************************/
@@ -354,7 +354,7 @@ void Pool::SetPoolIncrement (Local<String> property, Local<Value> value,
     msg = NJSMessages::getErrorMsg(errInvalidPool);
   else
     msg = NJSMessages::getErrorMsg(errReadOnly, "poolIncrement");
-  NJS_SET_EXCEPTION(msg.c_str(), msg.length());
+  NJS_SET_EXCEPTION(msg.c_str(), (int) msg.length());
 }
 
 /*****************************************************************************/
@@ -372,7 +372,7 @@ void Pool::SetPoolTimeout (Local<String> property, Local<Value> value,
     msg = NJSMessages::getErrorMsg(errInvalidPool);
   else
     msg = NJSMessages::getErrorMsg(errReadOnly, "poolTimeout");
-  NJS_SET_EXCEPTION(msg.c_str(), msg.length());
+  NJS_SET_EXCEPTION(msg.c_str(), (int) msg.length());
 }
 
 /*****************************************************************************/
@@ -390,7 +390,7 @@ void Pool::SetConnectionsOpen (Local<String> property, Local<Value> value,
     msg = NJSMessages::getErrorMsg(errInvalidPool);
   else
     msg = NJSMessages::getErrorMsg(errReadOnly, "connectionsOpen");
-  NJS_SET_EXCEPTION(msg.c_str(), msg.length());
+  NJS_SET_EXCEPTION(msg.c_str(), (int) msg.length());
 }
 
 /*****************************************************************************/
@@ -408,7 +408,7 @@ void Pool::SetConnectionsInUse (Local<String> property, Local<Value> value,
     msg = NJSMessages::getErrorMsg(errInvalidPool);
   else
     msg = NJSMessages::getErrorMsg(errReadOnly, "connectionsInUse");
-  NJS_SET_EXCEPTION(msg.c_str(), msg.length());
+  NJS_SET_EXCEPTION(msg.c_str(), (int) msg.length());
 }
 
 /*****************************************************************************/
@@ -426,7 +426,7 @@ void Pool::SetStmtCacheSize(Local<String> property, Local<Value> value,
     msg = NJSMessages::getErrorMsg(errInvalidPool);
   else
     msg = NJSMessages::getErrorMsg(errReadOnly, "stmtCacheSize");
-  NJS_SET_EXCEPTION(msg.c_str(), msg.length());
+  NJS_SET_EXCEPTION(msg.c_str(), (int) msg.length());
 }
 
 /*****************************************************************************/
